fix: pay back the dashboard lines by extraction, and record the new verb
check_source_size_budget: update_roadmap_progress.ts is past the 1500-line
ceiling, so the +69 lines the stub counts added were all charged. Paid back
by extraction rather than a baseline bump — stub_queue.ts takes the counter
and ticket_bundles.ts takes Bundle plus collect_bundles, both imported back
so behaviour is unchanged. 1565 -> 1454 lines, and the ratchet reads exactly
18512, its baseline.

The registry gained stubs:due, so evaluator-budgets.json records 107 -> 108
with a baseline_note naming what the verb is and why it is read-only, and
the committed measurement record is REGENERATED via
record_evaluator_measurements.mjs rather than hand-edited — which is what
its own remedy text demands and what the third preflight error was about.
