+++ internal/reports/secret-scanner-adversarial.json
@@ -1,6 +1,6 @@
 {
   "_doc": "Adversarial fixture measurement of the secret detector (scanText). Regenerated by src/scripts/_lib/secret_detector_adversarial.test.ts on every run; hard floor asserted there (recall 100% / FP 0% outside known_gaps).",
-  "generated": "2026-08-23",
+  "generated": "2026-08-26",
   "corpus_size": 31,
   "positives": 16,
   "negatives": 15,
